requirements.txt
uvicorn
langchain>=0.2
langchain_openai
langchain_community
streamlit
python-dotenv